backend/requirements.txt
aiohttp==3.13.2
alembic==1.17.2
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
boto3==1.40.74
certifi==2025.11.12
click==8.3.0
croniter==6.0.0
distro==1.9.0
fastapi==0.121.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.12.0
Mako==1.3.10
MarkupSafe==3.0.3
numpy==2.0.2
openai==2.8.0
opencv-python==4.12.0.88
packaging==25.0
pillow==12.0.0
psycopg2-binary==2.9.11
pydantic==2.12.4
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydub==0.25.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
pytz==2025.2
redis==7.0.1
requests==2.31.0
rq==2.6.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.49.3
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
wheel==0.45.1
rembg==2.0.61
librosa==0.10.0
scipy==1.13.1
